backend/requirements.txt
# backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.23.0
python-dotenv>=0.19.0
pydantic>=1.8.2
jinja2>=3.1.0
python-dotenv==1.1.0